Remove Syncers and related project tasks
* `remove_dirs` is not needed to be executed anymore on webs because
all the artifacts are saved on django storage backend now. So, this
task is being replaced by `clean_project_resources`.

* `move_files` is not needed because we don't need to sync files
anymore to the web servers. Everything is uploaded from build servers
to the django storage backend now.

* broadcast finalization steps are replaced by only a call to
`filefy`. The rest are not required anymore.

* `update_static_metadata` won't be needed because the redirection
done by the Flask app it will be replaced by El Proxito.

* Syncers are removed completely. This copying is happening on build
servers now and pushing the files to django storage backend.